Version increased to 0.9.8E1

Conflicts:
	dat/data.base
	include/artilist.h
	include/patchlevel.h
	src/apply.c
	src/artifact.c
	src/cmd.c
	src/mon.c
	src/monst.c
	src/potion.c
	src/tech.c
	src/timeout.c
	src/topten.c
	src/trap.c

+++ include/qt_win.h
@@ -300,8 +300,8 @@ class NetHackQtMapWindow : public QWidget, public NetHackQtWindow {
 	Q_OBJECT
 private:
 	NetHackQtClickBuffer& clicksink;
-	unsigned short glyph[ROWNO][COLNO];
-	unsigned short& Glyph(int x, int y) { return glyph[y][x]; }
+	unsigned /*short*/long glyph[ROWNO][COLNO];
+	unsigned /*short*/long& Glyph(int x, int y) { return glyph[y][x]; }
 	QPoint cursor;
 	BlackScrollView viewport;
 	QPixmap pet_annotation;
